feat: compute the full names of suites and tests (#31)
"full name" means the name prepended with all parent suite names.

This can be useful for cypress-grep and grepFilterSpecs.

At the moment grepFilterSpecs tests against the direct names
of tests and suites to determine if a spec should run.

Consider the following case:

```
    describe('parent', () => {
      describe('inner', () => {
        it('works a', () => {})
        it('works b', () => {})
      })
    })
```

and `grep=-parent`

The spec file cannot be filtered by grepFilterSpecs because
the inner suite, and the two tests don't have parent in the name. (The tests will be skipped during run-time).

With the full name, all suites and tests will be excluded and the spec file is never loaded.